Function confy::store_path

pub fn store_path<T: Serialize>(
    path: impl AsRef<Path>,
    cfg: T
) -> Result<(), ConfyError>

Save changes made to a configuration object at a specified path

This is an alternate version of store that allows the specification of an arbitrary path instead of a system one. For more information on errors and behavior, see store’s documentation.