confy
Function store_path_perms

pub fn store_path_perms<T: Serialize>(
    path: impl AsRef<Path>,
    cfg: T,
    perms: Permissions,
) -> Result<(), ConfyError>

Save changes made to a configuration object at a specified path

This is an alternate version of store_path that allows the specification of file permissions that must be set. For more information on errors and behavior, see store’s documentation.