confy
Function load_path

pub fn load_path<T: Serialize + DeserializeOwned + Default>(
    path: impl AsRef<Path>,
) -> Result<T, ConfyError>

Load an application configuration from a specified path.

A new configuration file is created with default values if none exists.

This is an alternate version of load that allows the specification of an arbitrary path instead of a system one. For more information on errors and behavior, see load’s documentation.